requirements.txt
transformers==4.25.1
numba==0.55.1
tqdm
julius
tabulate
einops
pydub
pypinyin
cn2an
jieba
soundfile
sounddevice
numpy
torch
torchaudio
librosa
argparse
pyyaml
